HBASE-25471 Move RegionScannerImpl out of HRegion (#2853)
