+++ hbase-server/src/test/java/org/apache/hadoop/hbase/regionserver/TestScannerHeartbeatMessages.java
@@ -53,7 +53,6 @@ import org.apache.hadoop.hbase.client.Table;
 import org.apache.hadoop.hbase.client.TableDescriptor;
 import org.apache.hadoop.hbase.filter.Filter;
 import org.apache.hadoop.hbase.filter.FilterBase;
-import org.apache.hadoop.hbase.regionserver.HRegion.RegionScannerImpl;
 import org.apache.hadoop.hbase.testclassification.LargeTests;
 import org.apache.hadoop.hbase.util.Bytes;
 import org.apache.hadoop.hbase.util.Threads;
@@ -525,7 +524,7 @@ public class TestScannerHeartbeatMessages {
   private static class HeartbeatReversedRegionScanner extends ReversedRegionScannerImpl {
     HeartbeatReversedRegionScanner(Scan scan, List<KeyValueScanner> additionalScanners,
         HRegion region) throws IOException {
-      super(scan, additionalScanners, region);
+      super(scan, additionalScanners, region, HConstants.NO_NONCE, HConstants.NO_NONCE);
     }
 
     @Override
@@ -554,7 +553,7 @@ public class TestScannerHeartbeatMessages {
   private static class HeartbeatRegionScanner extends RegionScannerImpl {
     HeartbeatRegionScanner(Scan scan, List<KeyValueScanner> additionalScanners, HRegion region)
         throws IOException {
-      region.super(scan, additionalScanners, region);
+      super(scan, additionalScanners, region, HConstants.NO_NONCE, HConstants.NO_NONCE);
     }
 
     @Override
